requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.29.0
pydantic==1.10.21
ping3==4.0.8
requests==2.31.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
bcrypt==4.0.1
